# optimize stringification