This reduces the number of operations in `bigDecimalToPlainString`. Every little bit counts. With this, toString / toFixed are on par with big.js for a few sample values, and decimal.js's toFixed. Somehow decimal.js toString is still WAY faster.

The tests seem to pass with this, but I didn't do a ton of checking to make sure it was OK.

```
BigDecimal.BigDecimal x 1,028,157 ops/sec ±0.40% (84 runs sampled)
Decimal#constructor x 434,684 ops/sec ±1.51% (96 runs sampled)
Big#constructor x 681,729 ops/sec ±0.76% (89 runs sampled)
Fastest is BigDecimal.BigDecimal
BigDecimal#toFixed x 645,770 ops/sec ±0.44% (95 runs sampled)
Decimal#toFixed x 621,232 ops/sec ±1.62% (92 runs sampled)
Big#toFixed x 657,078 ops/sec ±1.68% (96 runs sampled)
Fastest is Big#toFixed
BigDecimal#toString x 1,544,764 ops/sec ±2.10% (89 runs sampled)
Decimal#toString x 2,602,547 ops/sec ±0.85% (94 runs sampled)
Big#toString x 1,282,278 ops/sec ±2.24% (90 runs sampled)
Fastest is Decimal#toString
```

For the values

```js
let vals = [
  "98.987432987",
  "100000000000.00",
  "987344785.36",
  "1",
];
```