@imirkin 2023-08-20
I've just been doing the microbenchmarks on Node v18.16, because it's easier. The ultimate target is Chrome, I've been hoping that the results carry over.

Looks like the "+" isn't a big perf difference. ZERO_CHAR is a big jump:

Without ZERO_CHAR, from a few runs:
```
BigDecimal#toFixed x 618,175 ops/sec ±0.63% (95 runs sampled)
BigDecimal#toFixed x 623,340 ops/sec ±0.96% (91 runs sampled)
BigDecimal#toFixed x 609,253 ops/sec ±0.50% (94 runs sampled)
BigDecimal#toFixed x 629,028 ops/sec ±0.43% (94 runs sampled)
```

With ZERO_CHAR (and the +):
```
BigDecimal#toFixed x 648,698 ops/sec ±0.59% (97 runs sampled)
BigDecimal#toFixed x 643,930 ops/sec ±0.77% (94 runs sampled)
BigDecimal#toFixed x 638,103 ops/sec ±0.37% (96 runs sampled)
BigDecimal#toFixed x 649,597 ops/sec ±0.43% (93 runs sampled)
```

I've tried my best to make the setup stable for benchmarking, but apparently still not -- enabled performance governor, forced a single cpu for node running. Even though the results are slightly unstable, they do seem to indicate non-overlapping ranges.

Perhaps the node v8 is older than the Chrome one, and some of this is for naught, but I'm not sure how to get proper benchmarks for this stuff going in Chrome.