@imirkin 2023-08-20
BTW, you could still remove one layer of Math.max() and allow "zeros" to be negative. IME that's worth some improvement, but ... up to you.